chore: break down document into multiple widgets
