+++ lib/packages/ui/lib/src/widgets/media/document/document.dart
@@ -1,0 +1,3 @@
+export 'package:likeminds_chat_flutter_ui/src/widgets/media/document/preview.dart';
+export 'package:likeminds_chat_flutter_ui/src/widgets/media/document/thumbnail.dart';
+export 'package:likeminds_chat_flutter_ui/src/widgets/media/document/tile.dart';
