+++ lib/packages/ui/lib/src/widgets/widgets.dart
@@ -23,7 +23,7 @@ export 'shimmers/skeleton_chat_box.dart';
 
 export 'media/image.dart';
 export 'media/video.dart';
-export 'media/document.dart';
+export 'media/document/document.dart';
 
 export 'conversation/chat_bubble.dart';
